// src/colab/client/generated/schema-v1beta.d.ts
export interface components {
    schemas: {
// ...
        Key: {
            /** @description Required. Runtime variant. */
            variant: components["schemas"]["Variant"];
            /**
             * @description Required. Runtime accelerator model name (e.g., "T4", "L4", "A100").
             *     For CPU runtime specs, this field will be "NONE".
             */
            accelerator: string;
            /** @description Required. Runtime shape. */
            shape: components["schemas"]["Shape"];
        };
// ...
    };
// ...
}